forum tests moved into separate moduls
